requirements.txt
absl-py==0.9.0
astor==0.8.1
boto==2.49.0
boto3==1.11.3
botocore==1.14.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
gast==0.2.2
gensim==3.8.1
google-auth==1.10.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
jieba==0.42
jmespath==0.9.4
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
smart-open==1.9.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
